requirements.txt
# Медицинский ИИ Анализатор - Зависимости
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Telegram Bot
python-telegram-bot==20.7
python-telegram-bot[webhooks]==20.7

# AI/ML
openai==1.3.6
tiktoken==0.5.1

# Database
supabase==2.0.2
asyncpg==0.29.0
sqlalchemy==2.0.23

# File Processing
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
loguru==0.7.2
structlog==23.2.0

# HTTP & Async
httpx==0.25.2
aiofiles==23.2.0
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0

# Security
cryptography==41.0.8
bcrypt==4.1.2 